+++ tests/expand.rs
@@ -1,5 +1,5 @@
 use contextual::WithContext;
-use json_ld::{JsonLdProcessor, Loader, Print, RemoteDocumentReference, TryFromJson};
+use jsonld::{JsonLdProcessor, Loader, Print, RemoteDocumentReference, TryFromJson};
 use rdf_types::vocabulary::{IndexVocabulary, IriIndex, IriVocabularyMut};
 use static_iref::iri;
 
@@ -52,7 +52,7 @@ mod expand {
 		pub expand_context: Option<&'static Iri>,
 
 		#[iri("test:processingMode")]
-		pub processing_mode: Option<json_ld::ProcessingMode>,
+		pub processing_mode: Option<jsonld::ProcessingMode>,
 
 		#[iri("test:specVersion")]
 		pub spec_version: Option<&'static str>,
@@ -87,13 +87,13 @@ impl expand::Test {
 		}
 
 		let mut vocabulary: IndexVocabulary = IndexVocabulary::new();
-		let mut loader = json_ld::FsLoader::default();
+		let mut loader = jsonld::FsLoader::default();
 		loader.mount(
 			iri!("https://w3c.github.io/json-ld-api").to_owned(),
 			"tests/json-ld-api",
 		);
 
-		let mut options: json_ld::Options<IriIndex> = json_ld::Options::default();
+		let mut options: jsonld::Options<IriIndex> = jsonld::Options::default();
 		if let Some(p) = self.options.processing_mode {
 			options.processing_mode = p
 		}
@@ -121,7 +121,7 @@ impl expand::Test {
 					.unwrap()
 					.into_document();
 				let expected =
-					json_ld::ExpandedDocument::try_from_json_in(&mut vocabulary, expected).unwrap();
+					jsonld::ExpandedDocument::try_from_json_in(&mut vocabulary, expected).unwrap();
 
 				let success = expanded == expected;